+++ tests/utils/browser_paths.py
@@ -1,13 +1,15 @@
-import sys
 import os
+import sys
 
 from . import BrowserName, logger
+
 
 def get_username():
     try:
         return os.getlogin()
     except OSError:
         return os.environ.get('USERNAME') or os.environ.get('USER') or ''
+
 
 BIN_LOCATIONS = {
     BrowserName.CHROME: {
@@ -16,7 +18,8 @@ BIN_LOCATIONS = {
             r'C:\Program Files (x86)\Google\Chrome\Application\chrome.exe',
             r'C:\Program Files\Google\Chrome\Application\chrome.exe'
         ],
-        'macos': ['/Applications/Google Chrome.app/Contents/MacOS/Google Chrome'] # Not tested
+        # Not tested
+        'macos': ['/Applications/Google Chrome.app/Contents/MacOS/Google Chrome']
     },
     BrowserName.CHROMIUM: {
         'linux': ['/usr/bin/chromium', '/usr/bin/chromium-browser'],
@@ -24,7 +27,8 @@ BIN_LOCATIONS = {
             r'C:\Program Files (x86)\Chromium\Application\chrome.exe',
             r'C:\Program Files\Chromium\Application\chrome.exe'
         ],
-        'macos': ['/Applications/Chromium.app/Contents/MacOS/Chromium'] # Not tested
+        # Not tested
+        'macos': ['/Applications/Chromium.app/Contents/MacOS/Chromium']
     },
     BrowserName.BRAVE: {
         'linux': ['/usr/bin/brave', '/usr/bin/brave-browser'],
@@ -33,7 +37,8 @@ BIN_LOCATIONS = {
             r'C:\Program Files\BraveSoftware\Brave-Browser\Application\brave.exe',
             fr'C:\Users\{get_username()}\AppData\Local\BraveSoftware\Brave-Browser\Application\brave.exe'
         ],
-        'macos': ['/Applications/Brave Browser.app/Contents/MacOS/Brave Browser'] # Not tested
+        # Not tested
+        'macos': ['/Applications/Brave Browser.app/Contents/MacOS/Brave Browser']
     },
     BrowserName.EDGE: {
         'linux': ['/usr/bin/microsoft-edge-stable'],
@@ -41,7 +46,8 @@ BIN_LOCATIONS = {
             r'C:\Program Files (x86)\Microsoft\Edge\Application\msedge.exe',
             r'C:\Program Files\Microsoft\Edge\Application\msedge.exe'
         ],
-        'macos': ['/Applications/Microsoft Edge.app/Contents/MacOS/Microsoft Edge'] # Not tested
+        # Not tested
+        'macos': ['/Applications/Microsoft Edge.app/Contents/MacOS/Microsoft Edge']
     },
     BrowserName.FIREFOX: {
         'linux': ['/usr/bin/firefox'],
@@ -49,7 +55,18 @@ BIN_LOCATIONS = {
             r'C:\Program Files (x86)\Mozilla Firefox\firefox.exe',
             r'C:\Program Files\Mozilla Firefox\firefox.exe'
         ],
-        'macos': ['/Applications/Firefox.app/Contents/MacOS/firefox'] # Not tested
+        # Not tested
+        'macos': ['/Applications/Firefox.app/Contents/MacOS/firefox']
+    },
+    BrowserName.LIBREWOLF: {
+        'linux': ['/usr/bin/librewolf'],
+        'windows': [
+            r'C:\Program Files (x86)\LibreWolf\librewolf.exe',
+            r'C:\Program Files\LibreWolf\librewolf.exe',
+            fr'C:\Users\{get_username()}\AppData\Local\LibreWolf\librewolf.exe'
+        ],
+        # Not tested
+        'macos': ['/Applications/LibreWolf.app/Contents/MacOS/LibreWolf']
     },
     BrowserName.OPERA: {
         'linux': ['/usr/bin/opera'],
@@ -58,7 +75,7 @@ BIN_LOCATIONS = {
             r'C:\Program Files\Opera\opera.exe',
             fr'C:\Users\{get_username()}\AppData\Local\Programs\Opera\opera.exe'
         ],
-        'macos': ['/Applications/Opera.app/Contents/MacOS/Opera'] # Not tested
+        'macos': ['/Applications/Opera.app/Contents/MacOS/Opera']  # Not tested
     },
     BrowserName.OPERA_GX: {
         'linux': [],
@@ -67,7 +84,8 @@ BIN_LOCATIONS = {
             r'C:\Program Files\Opera GX\opera.exe',
             fr'C:\Users\{get_username()}\AppData\Local\Programs\Opera GX\opera.exe'
         ],
-        'macos': ['/Applications/Opera GX.app/Contents/MacOS/Opera GX'] # Not tested
+        # Not tested
+        'macos': ['/Applications/Opera GX.app/Contents/MacOS/Opera GX']
     },
     BrowserName.VIVALDI: {
         'linux': ['/usr/bin/vivaldi-stable'],
@@ -76,9 +94,11 @@ BIN_LOCATIONS = {
             r'C:\Program Files\Vivaldi\Application\vivaldi.exe',
             fr'C:\Users\{get_username()}\AppData\Local\Vivaldi\Application\vivaldi.exe'
         ],
-        'macos': ['/Applications/Vivaldi.app/Contents/MacOS/Vivaldi'] # Not tested
+        # Not tested
+        'macos': ['/Applications/Vivaldi.app/Contents/MacOS/Vivaldi']
     }
 }
+
 
 class BinaryLocation:
     def __init__(self, raise_not_found=False):
@@ -91,8 +111,8 @@ class BinaryLocation:
             self.__os = 'windows'
         else:
             raise ValueError('unsupported os')
-    
-    def get(self, browser:str) -> str:
+
+    def get(self, browser: str) -> str:
         for i in BIN_LOCATIONS[browser][self.__os]:
             if os.path.exists(i):
                 logger.info(f'found {browser} binary at: {i}')